--- src/components/ProfilePage.scss
@@ -1,15 +1,0 @@
-.profile-info {
-  display: flex;
-  //flex: 1;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 20px;
-}
-
-.link {
-  position: absolute;
-  bottom: 0px;
-  margin: 20px;
-  font-weight: bold;
-}
