+++ src/routes/account/links.ts
@@ -1,0 +1,3 @@
+export function linkToAccountPage() {
+  return "/account";
+}
